deduplicate code between NavigationModel and RoutingModel
