+++ libosmscout-client-qt/include/osmscout/NavigationModel.h
@@ -103,13 +103,7 @@ public slots:
   void onMaxAllowedSpeed(double maxAllowedSpeed);
 
 public:
-  enum Roles {
-    ShortDescriptionRole = Qt::UserRole + 1,
-    DescriptionRole = Qt::UserRole + 2,
-    TypeRole = Qt::UserRole + 3,
-    RoundaboutExitRole = Qt::UserRole + 4
-  };
-  Q_ENUM(Roles)
+  using Roles = RouteStep::Roles;
 
 public:
   NavigationModel();
