+++ libosmscout-client-qt/include/osmscout/RoutingModel.h
@@ -90,13 +90,7 @@ private:
   osmscout::BreakerRef  breaker;
 
 public:
-  enum Roles {
-    ShortDescriptionRole = Qt::UserRole + 1,
-    DescriptionRole = Qt::UserRole + 2,
-    TypeRole = Qt::UserRole + 3,
-    RoundaboutExitRole = Qt::UserRole + 4
-  };
-  Q_ENUM(Roles)
+  using Roles = RouteStep::Roles;
 
 public:
   RoutingListModel(QObject* parent = 0);
