+++ libosmscout-client-qt/src/osmscout/NavigationModel.cpp
@@ -192,22 +192,7 @@ QVariant NavigationModel::data(const QModelIndex &index, int role) const
   }
 
   RouteStep step=routeSteps[index.row()];
-
-  switch (role) {
-    case Qt::DisplayRole:
-    case ShortDescriptionRole:
-      return step.getShortDescription();
-    case DescriptionRole:
-      return step.getDescription();
-    case TypeRole:
-      return step.getType();
-    case RoundaboutExitRole:
-      return step.getRoundaboutExit();
-    default:
-      break;
-  }
-
-  return QVariant();
+  return step.data(role);
 }
 
 int NavigationModel::rowCount(const QModelIndex &/*parent*/) const
@@ -229,14 +214,7 @@ Qt::ItemFlags NavigationModel::flags(const QModelIndex &index) const
 
 QHash<int, QByteArray> NavigationModel::roleNames() const
 {
-  QHash<int, QByteArray> roles=QAbstractListModel::roleNames();
-
-  roles[ShortDescriptionRole] = "shortDescription";
-  roles[DescriptionRole] = "description";
-  roles[TypeRole] = "type";
-  roles[RoundaboutExitRole] = "roundaboutExit";
-
-  return roles;
+  return RouteStep::roleNames(QAbstractListModel::roleNames());
 }
